Add files via upload
This project showcases a CSS-rendered version of the Laos national flag. It was created as part of my journey to enhance my CSS skills and demonstrates the use of fundamental CSS properties for designing a flag without any images. The flag design includes three horizontal bands with accurate colors and proportions, reflecting the Laos national flag.